+++ python/openEMS/nf2ff.py
@@ -196,6 +196,12 @@ class nf2ff_results:
         cos_phi = np.cos(PHI)
         sin_phi = np.sin(PHI)
 
+        # the current format holds one compound complex dataset per frequency,
+        # stored (theta, phi). The legacy format, requested by the Octave and
+        # Matlab interface, splits the complex data into a real and an
+        # imaginary dataset and stores them (phi, theta).
+        self._legacy = 'f0_real' in h5_file['/nf2ff/E_theta/FD']
+
         self.E_theta = []
         self.E_phi   = []
         self.P_rad   = []
@@ -203,14 +209,24 @@ class nf2ff_results:
         self.E_cprh  = []
         self.E_cplh  = []
         for n in range(len(self.freq)):
-            E_theta = np.array(h5_file['/nf2ff/E_theta/FD/f{}_real'.format(n)]) + 1j*np.array(h5_file['/nf2ff/E_theta/FD/f{}_imag'.format(n)])
-            E_theta = np.swapaxes(E_theta, 0, 1)
-            E_phi   = np.array(h5_file['/nf2ff/E_phi/FD/f{}_real'.format(n)])   + 1j*np.array(h5_file['/nf2ff/E_phi/FD/f{}_imag'.format(n)])
-            E_phi   = np.swapaxes(E_phi, 0, 1)
-            self.P_rad  .append(np.swapaxes(np.array(h5_file['/nf2ff/P_rad/FD/f{}'.format(n)]), 0, 1))
+            E_theta = self._ReadFD(h5_file, 'E_theta', n)
+            E_phi   = self._ReadFD(h5_file, 'E_phi'  , n)
+            self.P_rad  .append(self._ReadFD(h5_file, 'P_rad', n))
 
             self.E_theta.append(E_theta)
             self.E_phi  .append(E_phi)
             self.E_norm .append(np.sqrt(np.abs(E_theta)**2 + np.abs(E_phi)**2))
             self.E_cprh .append((cos_phi+1j*sin_phi) * (E_theta+1j*E_phi)/np.sqrt(2.0))
             self.E_cplh .append((cos_phi-1j*sin_phi) * (E_theta-1j*E_phi)/np.sqrt(2.0))
+
+    def _ReadFD(self, h5_file, name, n):
+        """Read frequency index *n* of ``/nf2ff/<name>/FD`` as a (theta, phi) array."""
+        grp = h5_file['/nf2ff/{}/FD'.format(name)]
+        if not self._legacy:
+            # h5py maps the compound {r,i} type onto a native complex array
+            return np.array(grp['f{}'.format(n)])
+        if 'f{}_real'.format(n) in grp:
+            data = np.array(grp['f{}_real'.format(n)]) + 1j*np.array(grp['f{}_imag'.format(n)])
+        else:
+            data = np.array(grp['f{}'.format(n)])   # P_rad is real in both formats
+        return np.swapaxes(data, 0, 1)